下发盒子字段过滤、更新删除操作增加更新时间判断

+++ Server/box/box-api/src/main/java/com/wecon/box/api/AlarmCfgDataApi.java
@@ -61,7 +61,7 @@ public interface AlarmCfgDataApi {
 	 * @param ids
 	 * @return
      */
-	boolean batchDeleteByPlcId(List<Integer> ids);
+	boolean batchDeleteByPlcId(List<Long> ids);
 
 	/**
 	 * 批量删除报警配置数据根据alarm_cfg_id
