+++ Server/box/box-impl/src/main/java/com/wecon/box/impl/AlarmCfgDataImpl.java
@@ -111,13 +111,13 @@ public class AlarmCfgDataImpl implements AlarmCfgDataApi {
 	}
 
 	@Override
-	public boolean batchDeleteByPlcId(List<Integer> ids) {
+	public boolean batchDeleteByPlcId(List<Long> ids) {
 		if (null == ids || ids.size() == 0) {
 			return false;
 		}
 
 		StringBuilder idSb = new StringBuilder();
-		for (int id : ids) {
+		for (long id : ids) {
 			idSb.append(",").append(id);
 		}
 		String sql = "delete from alarm_cfg_data where alarm_cfg_id in(select alarmcfg_id from alarm_cfg where plc_id in("
